requirements.txt
PyYAML==6.0.2
dotenv==0.9.9
pandas==2.3.0
openai==1.86.0
torchvision==0.22.1
transformers==4.52.4
fastchat==0.1.0
psutil==7.0.0
accelerate==1.7.0
nltk==3.9.1
scikit-learn==1.6.1
datasets==3.6.0